@@ -3,7 +3,7 @@
   "author": {
     "name": "Call-em-all Engineering Team"
   },
-  "version": "0.7.2",
+  "version": "0.7.5",
   "description": "Material Design UI components built with React",
   "main": "./lib",
   "scripts": {
@@ -30,7 +30,7 @@
     "react-draggable2": "^0.5.1"
   },
   "peerDependencies": {
-    "react": ">=0.12",
+    "react": ">=0.13",
     "react-tap-event-plugin": ">=0.1.3"
   },
   "devDependencies": {
@@ -39,10 +39,10 @@
     "reactify": "^0.17.1",
     "react-router": "^0.13"
   },
-  "gitHead": "a2c859a57296916798042f1197f30fb2d3baf565",
-  "_id": "material-ui@0.7.2",
-  "_shasum": "6ad8311412c6474df7ea242d208c4e5b9cf324e6",
-  "_from": "material-ui@*",
+  "gitHead": "b3ce0b272e5468919d01b3d3ab6751a92f682b48",
+  "_id": "material-ui@0.7.5",
+  "_shasum": "808467575ce6e40f213a3434803ced6a159d3dc2",
+  "_from": "material-ui@>=0.7.5 <0.8.0",
   "_npmVersion": "2.7.4",
   "_nodeVersion": "0.12.1",
   "_npmUser": {
@@ -68,9 +68,10 @@
     }
   ],
   "dist": {
-    "shasum": "6ad8311412c6474df7ea242d208c4e5b9cf324e6",
-    "tarball": "http://registry.npmjs.org/material-ui/-/material-ui-0.7.2.tgz"
+    "shasum": "808467575ce6e40f213a3434803ced6a159d3dc2",
+    "tarball": "http://registry.npmjs.org/material-ui/-/material-ui-0.7.5.tgz"
   },
   "directories": {},
-  "_resolved": "https://registry.npmjs.org/material-ui/-/material-ui-0.7.2.tgz"
+  "_resolved": "https://registry.npmjs.org/material-ui/-/material-ui-0.7.5.tgz",
+  "readme": "ERROR: No README data found!"
 }
